// src/app/_dto/TblHFirstTax.dto.ts
export class TblHFirstTax {
    HFirstTax_SysID: number;
    HFirstTax_Code: string;
    HFirstTax_Name: string;
    HFirstTax_InputPerc: number;
    HFirstTax_OutputPerc: number;
    HFirstTax_Input_AccSysID: number;
    HFirstTax_Input_AccCode: number;
    HFirstTax_Input_AccName: string;
    HFirstTax_Output_AccSysID: number;
    HFirstTax_Output_AccCode: number;
    HFirstTax_Output_AccName: string;
    HFirstTax_Adv_AccSysID: number;
    HFirstTax_Adv_AccCode: number;
    HFirstTax_Adv_AccName: string;
    HFirstTax_Adv_TabNo: number;
    HFirstTaxGroup_SysID: number;
    HFirstTaxGroup_Code: string;
    HFirstTaxGroup_Name: string;
    HFirstTax_AcDe_Yn: string;
    HFirstTax_AcDe_SysID: number;
    HFirstTax_AcDe_Code: string;
    HFirstTax_AcDe_Name: string;

}